weather/requirements.txt
mcp
requests
python-dotenv 